refactor: remove model package
도멘인에 모델, 상수를 넣었다.
