+++ src/main/java/lotto/domain/LottoWithBonus.java
@@ -1,8 +1,8 @@
-package lotto.model;
+package lotto.domain;
 
 import java.util.List;
-import lotto.constants.ErrorCode;
-import lotto.constants.LottoConstants;
+import lotto.domain.constants.ErrorCode;
+import lotto.domain.constants.LottoConstants;
 
 public class LottoWithBonus {
