+++ src/main/java/lotto/domain/Money.java
@@ -1,7 +1,7 @@
-package lotto.model;
+package lotto.domain;
 
-import lotto.constants.ErrorCode;
-import lotto.constants.LottoConstants;
+import lotto.domain.constants.ErrorCode;
+import lotto.domain.constants.LottoConstants;
 
 public class Money {
