+++ src/main/java/lotto/domain/constants/LottoConstants.java
@@ -1,4 +1,4 @@
-package lotto.constants;
+package lotto.domain.constants;
 
 public class LottoConstants {
