+++ src/test/java/lotto/domain/CompareTest.java
@@ -1,13 +1,9 @@
 package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.model.Lotto;
-import lotto.model.LottoMachine;
-import lotto.model.LottoWithBonus;
 import org.junit.jupiter.api.Test;
 
 class CompareTest {
