+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
