+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,10 +1,10 @@
-package lotto.model;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import lotto.constants.ErrorCode;
+import lotto.domain.constants.ErrorCode;
 import org.junit.jupiter.api.Test;
 
 class MoneyTest {
